Implemented Settings menu and audiomixer
I implemented Settings menu which contains a full screen toggle, Screen resolution, Graphics, Music slider to set music volume and save the changes, as well as a SFX slider to set SFX volume and save the changes. The screen resolution and Graphics will only be visible once the game is built.