Implemented SaveLoad as well as Consumables
Implemented SaveAndLoad system for the game but it still has some bugs and issues which I will still be fixing and will get it to work. It's struggling with going back to mainmenu and reloading the game. It's also missing some features such as saving inventory data, potions, mana, healthbar, etc... My main focus was to get it to save gold coins initially and to create a new json file and save the data in the json file.